+++ apps/backend/platform/analytics/analytics.go
@@ -45,6 +45,32 @@ func FilterProperties(properties posthog.Properties) posthog.Properties {
 	return filteredProperties
 }
 
+func Identify(message posthog.Identify) {
+
+	if os.Getenv("DISABLE_ANALYTICS") == "true" {
+		return
+	}
+
+	apiKey := os.Getenv("POSTHOG_API_KEY")
+	if apiKey == "" {
+		log.Error().Msg("POSTHOG_API_KEY is not set")
+		return
+	}
+
+	client, err := getClient(apiKey)
+	if err != nil {
+		log.Error().Err(err).Msg("Error getting analytics client")
+	}
+
+	if os.Getenv("PIXELEYE_HOSTING") != "true" {
+		message.Properties = FilterProperties(message.Properties)
+	}
+
+	if err := client.Enqueue(message); err != nil {
+		log.Error().Err(err).Msg("Error sending analytics event")
+	}
+}
+
 func Track(message posthog.Capture) {
 
 	if os.Getenv("DISABLE_ANALYTICS") == "true" {