+++ integrations/cypress/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pixeleye/cypress
 
+## 0.6.4
+
+### Patch Changes
+
+- [`9c08dad`](https://github.com/pixeleye-io/pixeleye/commit/9c08dadb1115de7afdb2561dd63cd742b6b5c5c3) Thanks [@AlfieJones](https://github.com/AlfieJones)! - Exposing inner playwright scale option
+
+- Updated dependencies []:
+  - @pixeleye/cli-booth@0.6.5
+
 ## 0.6.3
 
 ### Patch Changes